Upgrade: omega_sign_in 插件 (适配 OmegaMatcherInterface)
