+++ src/plugins/omega_sign_in/command.py
@@ -9,7 +9,7 @@
 """
 
 from nonebot import get_driver
-from nonebot.adapters import Bot
+from nonebot.adapters.onebot.v11.bot import Bot as OneBotV11Bot
 from nonebot.adapters.onebot.v11.event import (
     PokeNotifyEvent as OneBotV11PokeNotifyEvent,
     GroupMessageEvent as OneBotV11GroupMessageEvent,
@@ -25,7 +25,7 @@ from src.params.handler import get_command_str_single_arg_parser_handler
 from src.params.rule import event_has_permission_level
 from src.service import enable_processor_state
 from .config import sign_in_config
-from .helpers import handle_generate_fortune_card, handle_generate_sign_in_card, handle_fix_sign_in
+from .handlers import handle_generate_fortune_card, handle_generate_sign_in_card, handle_fix_sign_in
 
 _COMMAND_START: set[str] = get_driver().config.command_start
 _DEFAULT_COMMAND_START: str = list(_COMMAND_START)[0] if _COMMAND_START else ''
@@ -42,9 +42,11 @@ sign_in = MatcherGroup(
 sign_in.on_command('签到', handlers=[handle_generate_sign_in_card])
 sign_in.on_command('老黄历', aliases={'好感度', '一言'}, handlers=[handle_generate_fortune_card])
 
+
 if sign_in_config.signin_plugin_enable_regex_matcher:
-    sign_in.on_regex(r'^签到$', handlers=[handle_generate_sign_in_card])
-    sign_in.on_regex(r'^(老黄历|今日(运势|人品)|一言|好感度|我的好感)$', handlers=[handle_generate_fortune_card])
+    sign_in.on_fullmatch(('签到',), handlers=[handle_generate_sign_in_card])
+    sign_in.on_fullmatch(('老黄历', '今日运势', '一言', '好感度',), handlers=[handle_generate_fortune_card])
+
 
 sign_in.on_command(
     '补签',
@@ -59,7 +61,7 @@ sign_in.on_command(
     priority=11,
     block=False
 ).handle()
-async def handle_poke_sign_in(bot: Bot, event: OneBotV11PokeNotifyEvent):
+async def handle_poke_sign_in(bot: OneBotV11Bot, event: OneBotV11PokeNotifyEvent):
     # 获取戳一戳用户身份
     if event.group_id is None:
         sender_data = await bot.get_stranger_info(user_id=event.user_id)
